update 2D hertzian contact examples
